requirements.txt
flask
kafka-python-ng
pyspark
